+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.0.4 - unreleased
+
+- SNES ROM container that automatically fixes the internal header's checksum
+- Cols picks how a tilemap's pages sit side by side when there are multiple pages
+- A few more small fixes
+
 ## v1.0.3 - 2026-09-20
 
 - Compress & Reshape plugin type, allows combining a compression & reshape plugin